+++ Cargo.toml
@@ -9,7 +9,6 @@ members = [
     "crates/relayer-rest",
     "crates/telemetry",
     "crates/chain-registry",
-    "crates/cardano-chain",
     "tools/integration-test",
     "tools/test-framework",
 ]
