--- crates/cardano-chain/Cargo.toml
@@ -1,51 +1,0 @@
-[package]
-name = "ibc-cardano-chain"
-version = "0.1.0"
-edition = "2021"
-license = "Apache-2.0"
-repository = "https://github.com/webisoftSoftware/hermes"
-authors = ["Webisoft <hello@webisoft.io>"]
-rust-version = "1.85.0"
-description = "Cardano chain implementation for IBC Relayer"
-
-[dependencies]
-# Hermes workspace dependencies
-ibc-relayer = { workspace = true }
-ibc-relayer-types = { workspace = true }
-ibc-proto = { workspace = true }
-
-# Standard workspace dependencies
-anyhow = { workspace = true, features = ["std"] }
-async-trait = { workspace = true }
-thiserror = { workspace = true }
-tracing = { workspace = true }
-tracing-subscriber = { workspace = true }
-humantime-serde = { workspace = true }
-tendermint = { workspace = true }
-tokio = { workspace = true, features = ["full"] }
-tonic = { workspace = true }
-prost = { workspace = true }
-serde = { workspace = true, features = ["derive"] }
-serde_json = { workspace = true }
-toml = { workspace = true }
-clap = { workspace = true, features = ["derive"] }
-hex = { workspace = true }
-hdpath = { workspace = true }
-
-# Cryptography
-ed25519-dalek = { workspace = true }
-tiny-bip39 = { version = "1.0", package = "tiny-bip39" }
-digest = { workspace = true }
-
-# Cardano-specific
-slip10 = "0.4"
-blake2 = "0.10"
-pallas-primitives = "0.30"
-pallas-codec = "0.30"
-lazy_static = "1.4"
-
-# Tendermint (required by ChainEndpoint trait)
-tendermint-rpc = { workspace = true }
-
-[dev-dependencies]
-
