--- crates/cardano-chain/src/config.rs
@@ -1,68 +1,0 @@
-//! Configuration for Cardano chain
-
-use ibc_relayer_types::core::ics24_host::identifier::ChainId;
-use serde::{Deserialize, Serialize};
-use std::path::PathBuf;
-use std::time::Duration;
-
-/// Minimal configuration for Cardano chain integration
-#[derive(Debug, Clone, Serialize, Deserialize)]
-#[serde(deny_unknown_fields)]
-pub struct CardanoChainConfig {
-    /// The chain's network identifier
-    pub id: ChainId,
-
-    /// Gateway gRPC endpoint URL
-    pub gateway_url: String,
-
-    /// Network ID (1 = mainnet, 0 = testnet)
-    pub network_id: u8,
-
-    /// Key name for signing
-    pub key_name: String,
-
-    /// Keystore type (test, file, etc.)
-    #[serde(default)]
-    pub key_store_type: ibc_relayer::keyring::Store,
-
-    /// Optional path to keystore folder
-    pub key_store_folder: Option<PathBuf>,
-
-    /// Account index for CIP-1852 derivation
-    #[serde(default)]
-    pub account: u32,
-
-    /// Maximum block time (for timeout calculations)
-    #[serde(default = "default_max_block_time", with = "humantime_serde")]
-    pub max_block_time: Duration,
-
-    /// Packet filter configuration
-    #[serde(default)]
-    pub packet_filter: ibc_relayer::config::PacketFilter,
-
-    /// Optional trust threshold (not used by Cardano but required by config interface)
-    #[serde(default)]
-    pub trust_threshold: Option<ibc_relayer_types::core::ics02_client::trust_threshold::TrustThreshold>,
-}
-
-fn default_max_block_time() -> Duration {
-    Duration::from_secs(30)
-}
-
-impl Default for CardanoChainConfig {
-    fn default() -> Self {
-        Self {
-            id: ChainId::from_string("cardano-test"),
-            gateway_url: "http://localhost:3001".to_string(),
-            network_id: 0,
-            key_name: "default".to_string(),
-            key_store_type: ibc_relayer::keyring::Store::Test,
-            key_store_folder: None,
-            account: 0,
-            max_block_time: default_max_block_time(),
-            packet_filter: ibc_relayer::config::PacketFilter::default(),
-            trust_threshold: None,
-        }
-    }
-}
-
