--- crates/cardano-chain/src/lib.rs
@@ -1,24 +1,0 @@
-//! Cardano chain implementation for IBC Relayer (Hermes)
-//!
-//! This crate provides a Cardano-specific implementation of the ChainEndpoint trait
-//! for the Hermes IBC relayer.
-
-pub mod any_conversions;
-pub mod chain_handle;
-pub mod config;
-pub mod endpoint;
-pub mod error;
-pub mod gateway_client;
-pub mod keyring;
-pub mod signer;
-pub mod signing_key_pair;
-pub mod types;
-
-// Re-export key types for convenience
-pub use config::CardanoChainConfig;
-pub use endpoint::CardanoChainEndpoint;
-pub use error::Error;
-pub use gateway_client::GatewayClient;
-pub use keyring::CardanoKeyring;
-pub use signing_key_pair::CardanoSigningKeyPair;
-
