+++ crates/relayer-cli/src/commands/keys/delete.rs
@@ -129,6 +129,7 @@ pub fn delete_key(config: &ChainConfig, key_name: &str) -> eyre::Result<()> {
             keyring.remove_key(key_name)?;
         }
         ChainConfig::Penumbra(_) => unimplemented!("no key support for penumbra"),
+        ChainConfig::Cardano(_) => unimplemented!("no key support for cardano"),
     }
     Ok(())
 }
@@ -156,6 +157,7 @@ pub fn delete_all_keys(config: &ChainConfig) -> eyre::Result<()> {
             }
         }
         ChainConfig::Penumbra(_) => unimplemented!("no key support for penumbra"),
+        ChainConfig::Cardano(_) => unimplemented!("no key support for cardano"),
     }
     Ok(())
 }
