+++ crates/relayer-cli/src/commands/listen.rs
@@ -208,6 +208,7 @@ fn subscribe(
             let subscription = monitor_tx.subscribe()?;
             Ok(subscription)
         }
+        ChainConfig::Cardano(_) => unimplemented!("event subscription not yet supported for cardano"),
     }
 }
 
@@ -218,6 +219,7 @@ fn detect_compatibility_mode(
     let rpc_addr = match config {
         ChainConfig::CosmosSdk(config) | ChainConfig::Namada(config) => config.rpc_addr.clone(),
         ChainConfig::Penumbra(config) => config.rpc_addr.clone(),
+        ChainConfig::Cardano(_) => unimplemented!("rpc_addr not yet supported for cardano"),
     };
 
     let client = HttpClient::builder(rpc_addr.try_into()?)
@@ -232,6 +234,7 @@ fn detect_compatibility_mode(
             let status = rt.block_on(client.status())?;
             penumbra::util::compat_mode_from_version(&config.compat_mode, status.node_info.version)?
         }
+        ChainConfig::Cardano(_) => unimplemented!("compat_mode not yet supported for cardano"),
     };
 
     Ok(compat_mode)
