+++ crates/relayer/src/chain/cardano/mod.rs
@@ -1,25 +1,27 @@
-//! Cardano chain configuration
+//! Cardano chain implementation for Hermes IBC relayer
 //!
-//! This module only contains configuration types. The actual CardanoChainEndpoint
-//! implementation lives in the ibc-cardano-chain crate to avoid circular dependencies.
-//!
-//! ## Architecture Note
-//!
-//! Due to Rust's module system, we cannot directly import CardanoChainEndpoint here because:
-//! - ibc-cardano-chain depends on ibc-relayer (for traits, errors, etc.)
-//! - ibc-relayer would depend on ibc-cardano-chain (to spawn chains)
-//! - This creates a circular dependency
-//!
-//! ## Integration Options
-//!
-//! 1. **Standalone Binary**: Run ibc-cardano-chain as a separate process
-//! 2. **Plugin System**: Future Hermes plugin architecture  
-//! 3. **Workspace Binary**: Create a top-level binary that depends on both crates
-//!
-//! For now, Cardano chains must be integrated at the application level, not within
-//! the ibc-relayer library itself.
+//! This module provides complete Cardano integration following the same pattern
+//! as Cosmos and Penumbra implementations in Hermes.
 
+pub mod any_conversions;
+pub mod chain_handle;
 pub mod config;
+pub mod endpoint;
+pub mod error;
+pub mod gateway_client;
+pub mod keyring;
+pub mod signer;
+pub mod signing_key_pair;
+pub mod types;
 
+// Re-export key types for convenience
 pub use config::CardanoConfig;
+pub use endpoint::CardanoChainEndpoint;
+pub use error::Error as CardanoError;
+pub use gateway_client::GatewayClient;
+pub use keyring::CardanoKeyring;
+pub use signing_key_pair::CardanoSigningKeyPair;
+
+// Type alias matching Cosmos/Penumbra pattern
+pub type CardanoChain = CardanoChainEndpoint;
