+++ crates/relayer/src/spawn.rs
@@ -7,8 +7,8 @@ use ibc_relayer_types::core::ics24_host::identifier::ChainId;
 
 use crate::{
     chain::{
-        cosmos::CosmosSdkChain, handle::ChainHandle, namada::NamadaChain,
-        penumbra::PenumbraChain, runtime::ChainRuntime,
+        cardano::CardanoChain, cosmos::CosmosSdkChain, handle::ChainHandle,
+        namada::NamadaChain, penumbra::PenumbraChain, runtime::ChainRuntime,
     },
     config::{ChainConfig, Config},
     error::Error as RelayerError,
@@ -87,17 +87,7 @@ pub fn spawn_chain_runtime_with_config<Handle: ChainHandle>(
         ChainConfig::CosmosSdk(_) => ChainRuntime::<CosmosSdkChain>::spawn(config, rt),
         ChainConfig::Namada(_) => ChainRuntime::<NamadaChain>::spawn(config, rt),
         ChainConfig::Penumbra(_) => ChainRuntime::<PenumbraChain>::spawn(config, rt),
-        ChainConfig::Cardano(_) => {
-            // Cardano chain spawning is handled by the standalone ibc-cardano-chain crate
-            // which implements ChainEndpoint directly. The circular dependency prevents
-            // us from importing it here. Users should use the cardano-chain binary or
-            // integrate ibc-cardano-chain directly in their application.
-            tracing::error!("Cardano chain spawning not yet integrated into Hermes spawn system");
-            tracing::info!("To use Cardano, run the ibc-cardano-chain binary separately");
-            return Err(SpawnError::relayer(crate::error::Error::config(
-                crate::config::Error::wrong_type(),
-            )));
-        }
+        ChainConfig::Cardano(_) => ChainRuntime::<CardanoChain>::spawn(config, rt),
     }
     .map_err(SpawnError::relayer)?;
