+++ tools/test-framework/src/util/interchain_security.rs
@@ -37,6 +37,7 @@ pub fn update_relayer_config_for_consumer_chain(config: &mut Config) {
             }
             ChainConfig::CosmosSdk(_) | ChainConfig::Namada(_) => {}
             ChainConfig::Penumbra(_) => { /* no-op Penumbra does not support CCV */ }
+            ChainConfig::Cardano(_) => { /* no-op Cardano does not support CCV */ }
         }
     }
 }
